Various Super I/O fixes and corrections.
 - VIA VT1211:
   - Add missing LDNs and respective code to handle them.
   - Add some TODOs for other stuff that needs fixing.
   - Use VT1211_SP1 instead of hardcoding the LDN number (2).
   - Fixup pnp_dev_info[] as per datasheet, but some TODOs remain.
   - Various coding style fixes and changes to u8/u16/etc.
 
 - Serverengines Pilot: Various coding style fixes and changes to u8/u16/etc.

Acked-by: Uwe Hermann <uwe@hermann-uwe.de>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@6030 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/superio/serverengines/pilot/pilot_early_init.c
@@ -29,7 +29,7 @@
  */
 static void pilot_early_init(device_t dev)
 {
-	unsigned port = dev >> 8;
+	u16 port = dev >> 8;
 
 	print_debug("Using port: ");
 	print_debug_hex16(port);
@@ -55,6 +55,7 @@ static void pilot_early_init(device_t dev)
 	pnp_set_enable(PNP_DEV(port, 0x3), 0);
 	pnp_exit_ext_func_mode(dev);
 */
+
 	pnp_enter_ext_func_mode(dev);
 	pnp_set_logical_device(PNP_DEV(port, 0x4));
 	pnp_exit_ext_func_mode(dev);
@@ -95,6 +96,7 @@ static void pilot_early_init(device_t dev)
 	pnp_enter_ext_func_mode(dev);
 	pnp_set_enable(PNP_DEV(port, 0x7), 0);
 	pnp_exit_ext_func_mode(dev);
+
 /*
 	pnp_enter_ext_func_mode(dev);
 	pnp_set_logical_device(PNP_DEV(port, 0x8));
